added v6

+++ src/main/java/dev/tsvinc/r2p/RefundPaymentRequest.java
@@ -5,8 +5,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 public record RefundPaymentRequest(
         @NotBlank @Size(max = 35) String endToEndId,
-        @NotNull @DecimalMin(value = "0.01") Double requestedAmount
+        @NotNull @DecimalMin(value = "0.01") BigDecimal requestedAmount
 ) {
 }
