+++ src/main/java/dev/tsvinc/r2p/SettlementDetails.java
@@ -8,13 +8,15 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import jakarta.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record SettlementDetails(
         @NotNull SettlementSystem settlementSystem,
         @NotBlank @Size(max = 36) String settlementSystemReferenceId,
         @Pattern(regexp = "RRN|TXNID|YPPID") String settlementSystemReferenceType,
-        @NotNull @DecimalMin(value = "0.01") Double settledAmount,
-        @NotBlank @Size(min = 3, max = 3/*, exactly = 3*/) String settledAmountCurrency,
+        @NotNull @DecimalMin(value = "0.01") BigDecimal settledAmount,
+        @NotBlank @Size(min = 3, max = 3) String settledAmountCurrency,
         @Size(max = 250) String message,
         @Size(max = 250) String unicodeEmoji,
         @Valid AccountDetails debtorAccountDetail,
